Move more types into rib_unit::rpki
